+++ src/managers/gpio.ts
@@ -25,13 +25,16 @@ OTHER DEALINGS IN THE SOFTWARE.
 
 import { IGPIOModule, IDigitalInput, IDigitalOutput } from 'core-io-types';
 import { Value, Mode } from 'abstract-io';
-import { getMode, setMode, getPeripheral, createInternalErrorMessage } from '../core';
+import { getMode, setMode, getPeripheral } from '../core';
 import { EventEmitter } from 'events';
+
+const READ_UPDATE_RATE = 18;
 
 export class GPIOManager {
 
   private module: IGPIOModule;
   private eventEmitter: EventEmitter;
+  private intervals: NodeJS.Timeout[] = [];
 
   constructor(gpioModule: IGPIOModule, globalEventEmitter: EventEmitter) {
     this.module = gpioModule;
@@ -74,41 +77,29 @@ export class GPIOManager {
       this.setInputMode(pin);
     }
 
-    // Use an arrow function so we can bind "this" properly
-    const addListener = () => {
+    let previousValue = -1;
+    const interval = setInterval(() => {
       const currentPeripheral = getPeripheral(pin);
       if (!currentPeripheral) {
-        throw new Error(createInternalErrorMessage(`peripheral is undefined even after setting the input mode`));
+        clearInterval(interval);
+        return;
       }
-
       switch (getMode(currentPeripheral)) {
         // Note: although we can only initiate this method in INPUT mode, we are supposed to continue
         // reporting values even if it's changed to OUTPUT mode
         case Mode.INPUT:
         case Mode.OUTPUT:
-          currentPeripheral.on('change', (value) => {
+          const value = (currentPeripheral as IDigitalInput).value;
+          if (value !== previousValue) {
+            previousValue = value;
             this.eventEmitter.emit(`digital-read-${pin}`, value);
             handler(value);
-          });
-
-          // Note: the peripheral instance will change if the pull resistor is changed or the mode is set to OUTPUT
-          // In both of these cases, we still want to emit digital-read-${pin} events
-          currentPeripheral.on('destroyed', () => setImmediate(addListener));
+          }
           break;
         default:
-          // Ignore all other modes, as the spec only calls for listening to input and output modes
+          clearInterval(interval);
       }
-    };
-    addListener();
-
-    // Force an initial "read" so that J5 knows the initial state of the pin
-    setImmediate(() => {
-      if (!peripheral) {
-        throw new Error(createInternalErrorMessage(`peripheral is undefined even after setting the input mode`));
-      }
-      const value = (peripheral as IDigitalInput).value;
-      this.eventEmitter.emit(`digital-read-${pin}`, value);
-      handler(value);
-    });
+    }, READ_UPDATE_RATE);
+    this.intervals.push(interval);
   }
 }